+++ src/app/auth/delete-account/page.tsx
@@ -1,4 +1,4 @@
-import DeleteAccountForm from "../../../components/auth/deleteAccountForm";
+import DeleteAccountForm from "../../../components/auth/DeleteAccountForm";
 
 export default function DeleteAccountPage() {
